requirements.txt
airtable==0.4.8
joblib==1.3.2
langchain==0.1.11
langchain_community==0.0.27
langchain_openai==0.0.8
python-dotenv==1.0.1
schedule==1.2.1
tweepy==4.14.0
faiss-cpu==1.8.0